+++ CHANGES.md
@@ -3,8 +3,8 @@
 LogTape changelog
 =================
 
-Version 0.13.0
---------------
+Version 1.0.0
+-------------
 
 To be released.